Add: overlay pattern to distinguish bios type
Not all my machines use UEFI, so to ensure repeatability, we need
to handle the case where we're using legacy bios, and so will be
using Grub.

This adds a little helper function that sets a let-bound variable
which we can use to decide between a different disk config based
on the presence or otherwise of efi.